% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtw-internal.R
\name{dtw-internal}
\alias{dtw-internal}
\alias{backtrack}
\alias{dtwpairdist}
\alias{globalCostMatrix}
\title{Internal dtw Functions}
\description{
Internal dtw functions
}
\details{
These are not to be called by the user. Frontend to the DTW package is the
\code{\link{dtw}} function.
}
\keyword{internal}
