\name{testmatrix}
\alias{testmatrix}
\alias{testq}
\alias{testt}

\docType{data}
\title{ Synthetic data sets (to be done) }
\description{
   Test sets for DTW 
}
\usage{data(testmatrix)}
\format{
  num [1:5] 0.861 0.711 0.535 0.376 0.761
}
\details{
  Will be improved
}
\source{
  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
  ~~ possibly secondary sources and usages ~~
}
\examples{
data(testmatrix)
## maybe str(testmatrix) ; plot(testmatrix) ...
}
\keyword{internal}
