\name{dtw-internal}
\alias{backtrack}
\alias{globalCostMatrix}
\alias{globalCostNative}

\title{Internal dtw Functions}
\description{
  Internal dtw functions
}
\details{
  These are not to be called by the user. Frontend to the
  DTW package is the \code{\link{dtw}} function.
}
\keyword{ internal }
